Moved all important parts of waveq_reverb into the wave_queue class and depricated the _reverb classes. Depricated the SharedPointerManager class. Included new envelope structure to store reverberation envelope data. Moved collision location from reflection model to wave_queue as volume interactions need access to that function. Added setter to the eigenverb_collection in reflection_model as a pass through in wave_queue.
