+++ waveq3d/reflection_model.h
@@ -181,24 +181,6 @@ class USML_DECLSPEC reflection_model
     bool surface_reflection( size_t de, size_t az ) ;
 
     /**
-     * Computes a refined location and direction at the point of collision.
-     * Uses a second order Taylor series around the current location to
-     * estimate these values.
-     *
-     * @param de            D/E angle index number.
-     * @param az            AZ angle index number.
-     * @param dtime         The distance (in time) from the "current"
-     *                      wavefront to the boundary collision.
-     * @param position      Refined position of the reflection (output).
-     * @param ndirection    Normalized direction at the point
-     *                      of reflection (output).
-     * @param speed         Speed of sound at the point of reflection (output).
-     */
-    void collision_location(
-        size_t de, size_t az, double dtime,
-        wposition1* position, wvector1* ndirection, double* speed ) const ;
-
-    /**
      * Re-initialize an individual ray after reflection.
      * Uses the position and reflected direction to initialize
      * a temporary 1x1 wavefront. Then, the position and direction
