--- waveq3d/waveq3d_reverb.h
@@ -1,15 +1,0 @@
-/**
- * @file waveq3d_reverb.h
- */
-
-#ifndef USML_WAVEQ3D_REVERB_H
-#define USML_WAVEQ3D_REVERB_H
-
-#include <usml/waveq3d/wave_queue.h>
-#include <usml/waveq3d/wave_front.h>
-#include <usml/waveq3d/wave_queue_reverb.h>
-#include <usml/eigenverb/eigenverb.h>
-#include <usml/eigenverb/eigenverb_monostatic.h>
-#include <usml/eigenverb/eigenverb_bistatic.h>
-
-#endif
